EIP1-8864: Moved the GenerateQualifyingAddressRequiredTemplatePreviewDto
