+++ src/main/kotlin/uk/gov/dluhc/notificationsapi/mapper/QualifyingAddressRequiredTemplatePreviewDtoMapper.kt
@@ -1,7 +1,7 @@
 package uk.gov.dluhc.notificationsapi.mapper
 
 import org.mapstruct.Mapper
-import uk.gov.dluhc.notificationsapi.dto.api.GenerateQualifyingAddressRequiredTemplatePreviewDto
+import uk.gov.dluhc.notificationsapi.dto.GenerateQualifyingAddressRequiredTemplatePreviewDto
 import uk.gov.dluhc.notificationsapi.models.GenerateQualifyingAddressRequiredTemplatePreviewRequest
 
 @Mapper(uses = [LanguageMapper::class, NotificationChannelMapper::class, SourceTypeMapper::class])
